requirements.txt
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.7.0
arabic-reshaper==2.1.3
asgiref==3.3.1
auditlog3==1.0.1
autopep8==1.5.5
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.5.2
django-celery==3.3.1
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-countries==7.0
django-crispy-forms==1.11.1
django-cron==0.5.1
django-crontab==0.7.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.0.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
kombu==3.0.37
oauthlib==3.0.1
pep8==1.7.1
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.1.0
pytz==2021.1
razorpay==1.2.0
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.4.1
stripe==2.27.0
toml==0.10.2
tzlocal==2.1
urllib3==1.24.2
virtualenv==20.4.2
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
